Improve ASM testing Removed the SpecFlowHelper, as this is now an extensions class. Added XML comments.
