+++ .github/workflows/build.yml
@@ -56,7 +56,7 @@ jobs:
 
     - name: Add Coverage PR Comment
       uses: marocchino/sticky-pull-request-comment@v2
-      if: github.event_name == 'pull_request'
+      if: github.event_name == 'pull_request' || github.event_name == 'pull_request_target'
       with:
         recreate: true
         path: code-coverage-results.md
