+++ src/Asm.Testing/ExceptionSteps.cs
@@ -1,10 +1,21 @@
 ﻿namespace Asm.Testing;
 
+/// <summary>
+/// Common steps for handling exceptions.
+/// </summary>
+/// <param name="context">Scenario context</param>
 [Binding]
 public class ExceptionSteps(ScenarioContext context)
 {
+    /// <summary>
+    /// The key used to store the exception in the context.
+    /// </summary>
     public const string ExceptionKey = nameof(ExceptionKey);
 
+    /// <summary>
+    /// Asserts an exception of the given type has been thrown.
+    /// </summary>
+    /// <param name="exceptionType">The fully qualified type name.</param>
     [Then(@"an exception of type '([^']*)' should be thrown")]
     [Then(@"an exception of type '([^']*)' is thrown")]
     [Then(@"an exception of type ""([^""]*)"" should be thrown")]
@@ -21,6 +32,10 @@ public class ExceptionSteps(ScenarioContext context)
         Assert.IsType(expected, actual);
     }
 
+    /// <summary>
+    /// Asserts the exception message is as expected.
+    /// </summary>
+    /// <param name="message">The message.</param>
     [Then(@"the exception message is '(.*)'")]
     [Then(@"the exception message should be '(.*)'")]
     [Then(@"the exception message is ""([^""]*)""")]
@@ -32,6 +47,13 @@ public class ExceptionSteps(ScenarioContext context)
         Assert.Equal(message.DecodeWhitespace(), ex.Message);
     }
 
+    /// <summary>
+    /// Asserts the argument exception parameter name is as expected.
+    /// </summary>
+    /// <remarks>
+    /// The exception thrown must have been of type <see cref="ArgumentException"/>.
+    /// </remarks>
+    /// <param name="parameterName">The name of the parameter.</param>
     [Then(@"the exception parameter name is '(.*)'")]
     [Then(@"the exception parameter name should be '(.*)'")]
     [Then(@"the exception parameter name is ""([^""]*)""")]
@@ -47,6 +69,9 @@ public class ExceptionSteps(ScenarioContext context)
         Assert.Equal(parameterName, ((ArgumentException)ex!).ParamName);
     }
 
+    /// <summary>
+    /// Asserts no exception has been thrown.
+    /// </summary>
     [Then(@"no exception is thrown")]
     public void ThenNoExceptionIsThrown()
     {
